changed BottomBar icons
