+++ example/lib/modules/wallet/components/core/wallet_page_components/wallet_page_build_nfts.dart
@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../../../../assets/icon_images/icons.dart';
+import '../../../../../../assets/icon_images/icons.dart';
 
+// ignore: must_be_immutable
 class BuildNft extends StatelessWidget {
   BuildNft({super.key});
+
   List<Map<dynamic, dynamic>> nfts = [
     {
       'icon': const AppIcon(
@@ -63,25 +65,23 @@ class BuildNft extends StatelessWidget {
             return Padding(
               padding: EdgeInsets.all(8.r),
               child: SizedBox(
-                child: Expanded(
-                  child: Stack(
-                    children: [
-                      nft['icon'],
-                      Positioned(
-                        bottom: screenWidth < 500 ? 7.h : 45.h,
-                        left: screenWidth < 500 ? 20.w : 35.w,
-                        child: Container(
-                          height: screenWidth < 500 ? 27.h : 50,
-                          width: screenWidth < 500 ? 90.w : 50.w,
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(8),
-                            color: Colors.black,
-                          ),
-                          child: Center(child: nft['name']),
+                child: Stack(
+                  children: [
+                    nft['icon'],
+                    Positioned(
+                      bottom: screenWidth < 500 ? 7.h : 45.h,
+                      left: screenWidth < 500 ? 20.w : 35.w,
+                      child: Container(
+                        height: screenWidth < 500 ? 27.h : 50,
+                        width: screenWidth < 500 ? 90.w : 50.w,
+                        decoration: BoxDecoration(
+                          borderRadius: BorderRadius.circular(8),
+                          color: Colors.black,
                         ),
-                      )
-                    ],
-                  ),
+                        child: Center(child: nft['name']),
+                      ),
+                    ),
+                  ],
                 ),
               ),
             );
